Move config utilities to its own file
